Refactor for better layout
