+++ src/check-failed-actions/check-failed.spec.ts
@@ -1,4 +1,4 @@
-// check-failed-actions/index.spec.ts
+// check-failed-actions/check-failed.spec.ts
 
 import { strict as assert } from 'node:assert';
 import os from 'node:os';
@@ -8,8 +8,7 @@ import http from 'node:http';
 import { v4 as uuid } from 'uuid';
 import getPort from 'get-port';
 
-import slackNock from './slack-nock.test';
-import { main } from './index';
+import checkFailed from './check-failed';
 
 describe('github', () => {
   const tmpFile = path.join(os.tmpdir(), uuid());
@@ -45,14 +44,12 @@ describe('github', () => {
   });
 
   it('Error - slack notified', async () => {
-    slackNock();
-
     process.env['GITHUB_EVENT_PATH'] = tmpFile;
     process.env['GITHUB_REPOSITORY'] = 'owner/repo';
     process.env['INPUT_FAILED'] = 'true';
     process.env['SLACK_FAILED_URL'] = `http://[::1]:${port}`;
 
-    await main();
+    await checkFailed();
     assert.equal(lastRequest.includes('owner/repo'), true);
     assert.equal(lastRequest.includes('*owner/repo - action failure*'), true);
   });
@@ -62,7 +59,7 @@ describe('github', () => {
     process.env['GITHUB_REPOSITORY'] = 'owner/repo';
     process.env['INPUT_FAILED'] = 'false';
 
-    await main();
+    await checkFailed();
     assert.ok(true);
   });
 });
